reorganization and restructuring of system exit logic (using an exception to handle user input is NOT good practice, and was temporary while this branch was being set up)
