+++ src/ui/OnReaderClosedListener.java
@@ -1,0 +1,5 @@
+package ui;
+
+public interface OnReaderClosedListener {
+    void onReaderClosed(); //callback
+}
